+++ src/main/java/com/nylas/Event.java
@@ -271,6 +271,13 @@ public class Event extends AccountOwnedModel implements JsonObject {
 	}
 
 	/**
+	 * Remove all metadata from the event
+	 */
+	public void clearMetadata() {
+		this.metadata.clear();
+	}
+
+	/**
 	 * Add one (or many) notifications to the event
 	 * @param notifications The notification(s) to append to the event's notification list
 	 */
@@ -283,12 +290,26 @@ public class Event extends AccountOwnedModel implements JsonObject {
 	}
 
 	/**
+	 * Remove all notifications from the event
+	 */
+	public void clearNotifications() {
+		this.notifications.clear();
+	}
+
+	/**
 	 * Add one (or many) participants to the event
 	 * @param participants The participant(s) to append to the event's participant list
 	 */
 	public void addParticipants(Participant... participants) {
 		this.participants.addAll(Arrays.asList(participants));
 		this.modifiedFields.put("participants", serializeParticipants());
+	}
+
+	/**
+	 * Remove all participants from the event
+	 */
+	public void clearParticipants() {
+		this.participants.clear();
 	}
 
 	/**
@@ -337,7 +358,7 @@ public class Event extends AccountOwnedModel implements JsonObject {
 
 	private List<Map<String, Object>> serializeParticipants() {
 		if(this.participants == null || this.participants.isEmpty()) {
-			return null;
+			return Collections.emptyList();
 		}
 
 		return this.participants.stream()